pyproject.toml
[project]
name = "riset-scylla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.3",
    "scylla-driver>=3.29.7",
]
